+++ src/main/java/ru/practicum/shareit/user/storage/UserStorage.java
@@ -7,13 +7,13 @@ import java.util.List;
 public interface UserStorage {
     List<User> getUsers();
 
-    User getUserById(Integer id);
+    User getUserById(Long id);
 
     User addUser(User user);
 
     User updateUser(User user);
 
-    void deleteUser(Integer id);
+    void deleteUser(Long id);
 
-    void checkUserAvailability(String operation, int id);
+    void checkUserAvailability(String operation, Long id);
 }
